Add type casting for x and z in transform_build_sample.py for consistency

+++ examples/transform_build_sample.py
@@ -7,8 +7,8 @@ __dirname__ = os.path.dirname(os.path.abspath(__file__))
 # Create a 3d sinusoidal wave pattern
 half_size = 0.5
 point_count = 50
-x = np.linspace(-half_size, half_size, point_count)
-z = np.linspace(-half_size, half_size, point_count)
+x = np.linspace(-half_size, half_size, point_count).astype(np.float32)
+z = np.linspace(-half_size, half_size, point_count).astype(np.float32)
 x, z = np.meshgrid(x, z)
 y = np.sin(np.sqrt(x**2 + z**2) * math.pi * 7) / 10
 positions = np.vstack((x.flatten(), y.flatten(), z.flatten())).T
